tests/wpt/tests/css/css-tables/parsing/table-layout-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module Test: parsing table-layout with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-table-layout">
<meta name="assert" content="table-layout supports the full grammar 'auto | fixed'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("table-layout", "auto");
test_valid_value("table-layout", "fixed");
</script>
</body>
</html>
